+++ src/index.ts
@@ -6799,6 +6799,12 @@ export default class Elysia<
 export { Elysia }
 
 export { t } from './type-system'
+export { validationDetail } from './type-system/utils'
+export type {
+	ElysiaTypeCustomError,
+	ElysiaTypeCustomErrorCallback
+} from './type-system/types'
+
 export { serializeCookie, Cookie, type CookieOptions } from './cookies'
 export type { Context, PreContext, ErrorContext } from './context'
 export {